feat(session replay): introduce standalone version of session replay package
