+++ packages/session-replay-browser/rollup.config.js
@@ -1,0 +1,6 @@
+import { iife, umd } from '../../scripts/build/rollup.config';
+
+iife.input = umd.input;
+iife.output.name = 'sessionReplay';
+
+export default [umd, iife];
