+++ packages/session-replay-browser/src/helpers.ts
