+++ packages/session-replay-browser/src/index.ts
@@ -1,0 +1,3 @@
+import sessionReplay from './session-replay-factory';
+export const { init, setSessionId, getSessionRecordingProperties, teardown } = sessionReplay;
+export { AmplitudeSessionReplay } from './typings/session-replay';
