+++ packages/session-replay-browser/test/helpers.test.ts
